correct multistream decompression EOF handling (fix #61)
when the input buffer has an end-of-stream marker right at the
beginning, decompress() will return StreamEnd and total_in will not
advance. We cannot return Ok(read) as this would signal EOF. Instead, we
rely on the next loop iteration to really return EOF when the input
buffer did not fill again.