write::BzDecoder: Fix infinite loop on drop when no data is read or written
If writes to the BzDecoder end before it has been fed the entire stream
and the output buffer has already been flushed to the wrapped writer,
then the Drop implementation will loop forever. On drop, try_finish() is
called, which repeatedly tries to write() until BZ_STREAM_END is
returned or an error occurs, but neither scenario happens if there's no
input nor output data to process.

This commit makes try_finish() return an UnexpectedEof error in this
scenario.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>